02_data_ingestion/requirements.txt
opendatasets
kaggle
apache-airflow-providers-google
pyspark
apache-airflow-providers-apache-spark
gcsfs